Add files via upload
Customer personality analysis helps a business to modify its product based on its target
customers from different types of customer segments. For example, instead of spending money
to market a new product to every customer in the company’s database, a company can analyze
which customer segment is most likely to buy the product and then market the product only on
that particular segment.

Target:   Need to perform clustering to summarize customer segments.