Add files via upload
A combined-cycle power plant comprises gas turbines, steam turbines, and heat
recovery steam generators. In this type of plant, the electricity is generated by gas and
steam turbines combined in one cycle. Then, it is transferred from one turbine to
another. We have to model the energy generated as a function of exhaust vacuum and
ambient variables and use that model to improve the plant's performance.